Chore: separate `recovery` queries into own file
