+++ src/data/graphql/queries.graphql
@@ -265,75 +265,6 @@ query EventsForMotion($motionId: Int!, $colonyAddress: String!) {
   }
 }
 
-query RecoveryEventsForSession($blockNumber: Int!, $colonyAddress: String!) {
-  recoveryEventsForSession(blockNumber: $blockNumber, colonyAddress: $colonyAddress) @client {
-    type
-    name
-    values
-    createdAt
-    emmitedBy
-    blockNumber
-    transactionHash
-  }
-}
-
-query RecoverySystemMessagesForSession($blockNumber: Int!, $colonyAddress: String!) {
-  recoverySystemMessagesForSession(blockNumber: $blockNumber, colonyAddress: $colonyAddress) @client {
-    type
-    name
-    createdAt
-  }
-}
-
-query GetRecoveryStorageSlot($colonyAddress: String!, $storageSlot: String!) {
-  getRecoveryStorageSlot(colonyAddress: $colonyAddress, storageSlot: $storageSlot) @client
-}
-
-query RecoveryRolesUsers($colonyAddress: String!, $endBlockNumber: Int) {
-  recoveryRolesUsers(colonyAddress: $colonyAddress, endBlockNumber: $endBlockNumber) @client {
-    id
-    profile {
-      avatarHash
-      displayName
-      username
-      walletAddress
-    }
-  }
-}
-
-query RecoveryRolesAndApprovalsForSession($blockNumber: Int!, $colonyAddress: String!) {
-  recoveryRolesAndApprovalsForSession(blockNumber: $blockNumber, colonyAddress: $colonyAddress) @client {
-    id
-    profile {
-      avatarHash
-      displayName
-      username
-      walletAddress
-    }
-    approvedRecoveryExit
-  }
-}
-
-query GetRecoveryRequiredApprovals($blockNumber: Int!,$colonyAddress: String!) {
-  getRecoveryRequiredApprovals(blockNumber: $blockNumber, colonyAddress: $colonyAddress) @client
-}
-
-query RecoveryAllEnteredEvents($colonyAddress: String!, $currentBlock: Int!) {
-  recoveryAllEnteredEvents(colonyAddress: $colonyAddress, currentBlock: $currentBlock) @client {
-    type
-    name
-    values
-    createdAt
-    emmitedBy
-    blockNumber
-    transactionHash
-  }
-}
-
-query LegacyNumberOfRecoveryRoles($colonyAddress: String!) {
-  legacyNumberOfRecoveryRoles(colonyAddress: $colonyAddress) @client
-}
-
 # Motions
 #
 #
