+++ src/data/graphql/typeDefs.ts
@@ -235,12 +235,6 @@ export default gql`
     processedColonies: [ProcessedColony!]!
   }
 
-  type UsersAndRecoveryApprovals {
-    id: String!
-    profile: UserProfile!
-    approvedRecoveryExit: Boolean!
-  }
-
   type ActionThatNeedsAttention {
     transactionHash: String!
     needsAction: Boolean!
@@ -339,32 +333,6 @@ export default gql`
       walletAddress: String!
     ): [ActionThatNeedsAttention]!
     eventsForMotion(motionId: Int!, colonyAddress: String!): [ParsedEvent!]!
-    recoveryEventsForSession(
-      blockNumber: Int!
-      colonyAddress: String!
-    ): [ParsedEvent!]!
-    recoverySystemMessagesForSession(
-      blockNumber: Int!
-      colonyAddress: String!
-    ): [SystemMessage!]!
-    recoveryRolesUsers(colonyAddress: String!, endBlockNumber: Int): [User!]!
-    getRecoveryStorageSlot(
-      colonyAddress: String!
-      storageSlot: String!
-    ): String!
-    recoveryRolesAndApprovalsForSession(
-      blockNumber: Int!
-      colonyAddress: String!
-    ): [UsersAndRecoveryApprovals!]!
-    getRecoveryRequiredApprovals(
-      blockNumber: Int!
-      colonyAddress: String!
-    ): Int!
-    recoveryAllEnteredEvents(
-      colonyAddress: String!
-      currentBlock: Int!
-    ): [ParsedEvent!]!
-    legacyNumberOfRecoveryRoles(colonyAddress: String!): Int!
     motionTimeoutPeriods(
       motionId: Int!
       colonyAddress: String!
